+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/recapadoras/RecapadoraDaoImpl.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.frota.pneu.recapadoras;
 
-import br.com.zalf.prolog.webservice.commons.util.Now;
+import br.com.zalf.prolog.webservice.commons.util.date.Now;
 import br.com.zalf.prolog.webservice.database.DatabaseConnection;
 import org.jetbrains.annotations.NotNull;