+++ CHANGELOG.md
@@ -1,14 +1,25 @@
 Change Log
 ==========
 
+<a name="v0.0.71"></a>
+## Version [v0.0.71](https://github.com/luizfp/PrologWeb/compare/v0.0.70...v0.0.71) (release-date) [unreleased]
+
+### Bug Fixes
+* Corrige cálculo das horas noturnas no relatório de folha de ponto do controle de jornada
+
 <a name="v0.0.70"></a>
-## Version [v0.0.70](https://github.com/luizfp/PrologWeb/compare/v0.0.69...v0.0.70) (release-date) [unreleased]
+## Version [v0.0.70](https://github.com/luizfp/PrologWeb/compare/v0.0.69...v0.0.70) (2018-07-29)
+
+### Features
+* Cria relatório com o tempo total por tipo de marcação no controle de jornada
+* Cálcula horas noturnas no relatório de folha de ponto
 
 <a name="v0.0.69"></a>
 ## Version [v0.0.69](https://github.com/luizfp/PrologWeb/compare/v0.0.68...v0.0.69) (2018-07-19)
 
 ### Features
 * Cria validator para o insert do Modelo de Quiz
+* Cria relatório para mostrar o total de tempo gasto em cada tipo de intervalo por cada colaborador
 
 ### Refactors
 * Adiciona permissão de movimentação na busca de uma recapadora específica
@@ -58,7 +69,6 @@ Change Log
 ### Features
 * Cria back end para salvar mensagens de contato de possíveis clientes
 * Ativa método de listagem de marcações de um colaborador
-* Cria relatório para mostrar o total de tempo gasto em cada tipo de intervalo por cada colaborador
 
 ### Refactor
 * Altera componente da Dashboard para mostrar apenas os veículos ativos e não apenas os veículos com pneus vinculados